test: update Validation test suite for record.py and table.py
