+++ tests/__mocks__.py
@@ -53,93 +53,3 @@ MOCK_STAKEHOLDERS = {
         "user_float_5": 1.0,
     },
 }
-
-MOCK_VALIDATIONS = {
-    1: {
-        "acceptable_maximum": 30.0,
-        "acceptable_mean": 20.0,
-        "acceptable_minimum": 10.0,
-        "acceptable_variance": 0.0,
-        "confidence": 95.0,
-        "cost_average": 0.0,
-        "cost_ll": 0.0,
-        "cost_maximum": 0.0,
-        "cost_mean": 0.0,
-        "cost_minimum": 0.0,
-        "cost_ul": 0.0,
-        "cost_variance": 0.0,
-        "date_end": date.today() + timedelta(days=30),
-        "date_start": date.today(),
-        "description": "",
-        "measurement_unit": 0,
-        "name": "PRF-0001",
-        "status": 0.0,
-        "task_type": 0,
-        "task_specification": "",
-        "time_average": 0.0,
-        "time_ll": 0.0,
-        "time_maximum": 0.0,
-        "time_mean": 0.0,
-        "time_minimum": 0.0,
-        "time_ul": 0.0,
-        "time_variance": 0.0,
-    },
-    2: {
-        "acceptable_maximum": 30.0,
-        "acceptable_mean": 20.0,
-        "acceptable_minimum": 10.0,
-        "acceptable_variance": 0.0,
-        "confidence": 95.0,
-        "cost_average": 0.0,
-        "cost_ll": 0.0,
-        "cost_maximum": 0.0,
-        "cost_mean": 0.0,
-        "cost_minimum": 0.0,
-        "cost_ul": 0.0,
-        "cost_variance": 0.0,
-        "date_end": date.today() + timedelta(days=20),
-        "date_start": date.today() - timedelta(days=10),
-        "description": "",
-        "measurement_unit": 0,
-        "name": "",
-        "status": 0.0,
-        "task_type": 5,
-        "task_specification": "",
-        "time_average": 0.0,
-        "time_ll": 0.0,
-        "time_maximum": 0.0,
-        "time_mean": 0.0,
-        "time_minimum": 0.0,
-        "time_ul": 0.0,
-        "time_variance": 0.0,
-    },
-    3: {
-        "acceptable_maximum": 30.0,
-        "acceptable_mean": 20.0,
-        "acceptable_minimum": 10.0,
-        "acceptable_variance": 0.0,
-        "confidence": 95.0,
-        "cost_average": 0.0,
-        "cost_ll": 0.0,
-        "cost_maximum": 0.0,
-        "cost_mean": 0.0,
-        "cost_minimum": 0.0,
-        "cost_ul": 0.0,
-        "cost_variance": 0.0,
-        "date_end": date.today() + timedelta(days=30),
-        "date_start": date.today(),
-        "description": "",
-        "measurement_unit": 0,
-        "name": "",
-        "status": 0.0,
-        "task_type": 5,
-        "task_specification": "",
-        "time_average": 20.0,
-        "time_ll": 19.0,
-        "time_maximum": 40.0,
-        "time_mean": 34.0,
-        "time_minimum": 12.0,
-        "time_ul": 49.0,
-        "time_variance": 0.0,
-    },
-}
